+++ resources/views/livewire/author-login-form.blade.php
@@ -37,7 +37,7 @@
                     <!-- <a href="#" class="link-secondary" id="togglePassword" data-bs-toggle="tooltip" aria-label="Show password" data-bs-original-title="Show password">Download SVG icon from http://tabler-icons.io/i/eye -->
                     <!-- <svg xmlns="http://www.w3.org/2000/svg" class="icon" width="24" height="24" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round"><path stroke="none" d="M0 0h24v24H0z" fill="none"></path><path d="M10 12a2 2 0 1 0 4 0a2 2 0 0 0 -4 0"></path><path d="M21 12c-2.4 4 -5.4 6 -9 6c-3.6 0 -6.6 -2 -9 -6c2.4 -4 5.4 -6 9 -6c3.6 0 6.6 2 9 6"></path></svg> -->
                     <!-- </a> -->
-                    <i class="far fa-eye" id="togglePassword"></i>
+                    <i class="ti ti-eye" id="togglePassword"></i>
                 </span>
                 </div>
                 @error('password')
@@ -67,8 +67,8 @@
             // toggle the type attribute
             const type = password.getAttribute('type') === 'password' ? 'text' : 'password';
             password.setAttribute('type', type);
-            // toggle the eye slash icon
-            this.classList.toggle('fa-eye-slash');
+            // toggle the eye slash icon  - using tabler icons
+            this.classList.toggle('ti-eye-off');
         });
     </script>
 @endpush