+++ resources/views/livewire/top-header.blade.php
@@ -142,7 +142,7 @@
                 </li>
                 <li class="nav-item">
                   <a class="nav-link" href="./" >
-                    <span class="nav-link-icon d-md-none d-lg-inline-block"><!-- Download SVG icon from https://tabler.io/icons/icon/categoryy -->
+                    <span class="nav-link-icon d-md-none d-lg-inline-block"><!-- Download SVG icon from https://tabler.io/icons/icon/category -->
                       <svg  xmlns="http://www.w3.org/2000/svg"  width="24"  height="24"  viewBox="0 0 24 24"  fill="none"  stroke="currentColor"  stroke-width="2"  stroke-linecap="round"  stroke-linejoin="round"  class="icon icon-tabler icons-tabler-outline icon-tabler-category"><path stroke="none" d="M0 0h24v24H0z" fill="none"/><path d="M4 4h6v6h-6z" /><path d="M14 4h6v6h-6z" /><path d="M4 14h6v6h-6z" /><path d="M17 17m-3 0a3 3 0 1 0 6 0a3 3 0 1 0 -6 0" /></svg>
                     </span>
                     <span class="nav-link-title">
@@ -150,6 +150,19 @@
                     </span>
                   </a>
                 </li>
+                @if(auth()->user()->type == 1)  <!-- {-- only available to admin users --} -->
+                  <li class="nav-item">
+                    <a class="nav-link" href="{{ route('author.authors') }}" >
+                      <span class="nav-link-icon d-md-none d-lg-inline-block"><!-- Download SVG icon from https://tabler.io/icons/icon/category -->
+                        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="icon icon-tabler icons-tabler-outline icon-tabler-users">  <path stroke="none" d="M0 0h24v24H0z" fill="none" />  <path d="M9 7m-4 0a4 4 0 1 0 8 0a4 4 0 1 0 -8 0" />  <path d="M3 21v-2a4 4 0 0 1 4 -4h4a4 4 0 0 1 4 4v2" />  <path d="M16 3.13a4 4 0 0 1 0 7.75" />  <path d="M21 21v-2a4 4 0 0 0 -3 -3.85" /></svg>
+                      </span>
+                      <span class="nav-link-title">
+                        {{ __('Authors') }}
+                      </span>
+                    </a>
+                  </li>
+                @endif
+
                 <li class="nav-item dropdown">
                   <a class="nav-link dropdown-toggle" href="#navbar-base" data-bs-toggle="dropdown" data-bs-auto-close="outside" role="button" aria-expanded="false" >
                     <span class="nav-link-icon d-md-none d-lg-inline-block"><!-- Download SVG icon from http://tabler-icons.io/i/notes -->