+++ resources/views/new-author-email-template.blade.php
@@ -1,0 +1,14 @@
+<br>
+{{ $txtMail['01'] }}, <b>{{ $name }} </b> <br> <br>
+{{ $txtMail['02'] }}.
+{{ $txtMail['03'] }}:
+<br>
+<b>{{ $txtMail['04'] }}</b>: {{ $username }}<br>
+<b>{{ $txtMail['05'] }}</b>: {{ $email }}<br>
+<b>{{ $txtMail['06'] }}</b>: {{ $password }}<br>
+<br><br>
+<a href="{{ $url }}">{{ $txtMail['07'] }}</a>
+<br>{{ $txtMail['08'] }}<br>
+<b></b>: {{ $txtMail['09'] }}.
+<br><br>
+{{ $txtMail['10'] }}.