+++ routes/author.php
@@ -19,6 +19,7 @@ Route::prefix('author')->name('author.')->group(function(){
         Route::view('/settings','back.pages.settings')->name('settings');
         Route::post('/change-blog-logo',[AuthorController::class, 'changeBlogPic'])->name('change-blog-logo');
         Route::post('/change-blog-favicon',[AuthorController::class, 'changeBlogPic'])->name('change-blog-favicon');
+        Route::view('/authors','back.pages.authors')->name('authors');
     });
 
 });