+++ CHANGELOG.md
@@ -181,3 +181,20 @@
 - resources\views\back\layouts     [`auth-layout.blade.php`](./resources/views/back/layouts/auth-layout.blade.php)
 - resources\views\livewire            [`author-login-form`](./resources/views/livewire/author-login-form.blade.php)
 - resources\views\livewire            [`author-reset-form`](./resources/views/livewire/author-reset-form.blade.php)
+
+> 22:  Toggle eye icon to tabler.icons + Add Author to blog
+
+>> 22.1 Toggle eye icon to tabler.icons
+
+- resources\views\back\layouts      [`auth-layout.blade.php`](./resources/views/back/layouts/auth-layout.blade.php)
+- resources\views\livewire      [`author-login-form.blade.php`](./resources/views/livewire/author-login-form.blade.php)
+- resources\views\livewire      [`author-reset-form.blade.php`](./resources/views/livewire/author-reset-form.blade.php)
+
+>> 22.2 Add Author to blog
+- resources\views\livewire     [`top-header.blade.php`](./resources/views/livewire/top-header.blade.php) 
+- lang    [`es.json`](./lang/es.json)    [`fr.json`](./lang/fr.json)
+- routes  [`author.php`](./routes/author.php)
+- resources\views\back\pages      [`authors.blade.php`](./resources/views/back/pages/authors.php)
+- CLASS: app\Livewire\Authors.php    [`Authors.php`](./app/Livewire/Authors.php)
+- VIEW: resources\views\livewire     [`authors.blade.php`](./resources/views/livewire)
+- resources\views     [`new-author-email-template.blade.php`](./resources/views/new-author-email-template.blade.php)